+++ src/main/java/com/hanghae/naegahama/repository/AnswerLikeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface AnswerLikeRepository extends JpaRepository<AnswerLike, Long> {
-    Long countByAnswer(Answer answer);
+    int countByAnswer(Answer answer);
     void deleteByAnswer(Answer answer);
     Optional<AnswerLike> findByUserAndAnswer(User user, Answer answer);
